bugfix: activePeriod check off-by-one error

+++ src/infosme/testShowSizes.cpp
@@ -25,14 +25,14 @@ int main()
 {
     smsc::infosme::CsvStore::CsvFile::Record rec;
     rec.msg.abonent = ".1.1.79137654079";
-    printf("sizeof(std::string)=%u\n",sizeof(std::string));
-    printf("sizeof(FixedLengthString)=%u\n",sizeof(smsc::core::buffers::FixedLengthString<16>));
-    printf("sizeof(CsvStore::Rec)=%u\n",sizeof(rec));
-    printf("sizeof(Message)=%u\n",sizeof(rec.msg));
+    printf("sizeof(std::string)=%u\n",unsigned(sizeof(std::string)));
+    printf("sizeof(FixedLengthString)=%u\n",unsigned(sizeof(smsc::core::buffers::FixedLengthString<16>)));
+    printf("sizeof(CsvStore::Rec)=%u\n",unsigned(sizeof(rec)));
+    printf("sizeof(Message)=%u\n",unsigned(sizeof(rec.msg)));
     printf("Please add 5 + %u + len(userData)+1 + len(message)+1\n",rec.msg.abonent.size()+1);
     printf("\n");
-    printf("sizeof(TestMsg)=%u\n",sizeof(TestMsg));
-    printf("sizeof(TestRec)=%u\n",sizeof(TestRec));
+    printf("sizeof(TestMsg)=%u\n",unsigned(sizeof(TestMsg)));
+    printf("sizeof(TestRec)=%u\n",unsigned(sizeof(TestRec)));
     printf("Please add len(userData)+1\n");
     return 0;
 }
